refactor(modules-yaml): split tests; pin chrono patch; angle-bracket URLs
Three concerns addressed in one commit:

- Split `tests/index.rs` (~525 lines) along the natural seam between
  upstream ports and pacquet-side additions, then split the pacquet
  tail by data source. Three files now:

    - `tests/index.rs` keeps the five direct ports of upstream's
      `installing/modules-yaml/test/index.ts` and preserves the
      `index.ts` ↔ `index.rs` filename mapping.
    - `tests/real_fs.rs` holds the four pacquet-side tests that
      exercise behavior branches via real `tempfile::tempdir()` writes
      and reads (absolute `virtualStoreDir`, `skipped` sort, null
      `publicHoistPattern`, `DepPath` transparent serialization).
    - `tests/fakes.rs` holds the eight pacquet-side tests that drive
      behavior through the capability-trait DI fakes (the five I/O
      error paths, `LayoutVersion` rejection, `ignoredBuilds` dedup,
      `prunedAt` clock fill-in).

  `manifest_from_json` is duplicated in `tests/index.rs` and
  `tests/real_fs.rs` (each file uses it independently); `tests/fakes.rs`
  doesn't need it. The recent pipe-trait refactors on
  `read_legacy_shamefully_hoist_*`,
  `read_preserves_absolute_virtual_store_dir`, and the round-trip test
  are preserved verbatim and `read_empty_modules_manifest_returns_none`
  picks up the same chain shape for consistency.

- Pin the workspace `chrono` dependency to `0.4.44` (the version the
  lockfile resolved). Other workspace deps follow the same patch-level
  pinning convention (`tar = "0.4.45"`, `dunce = "1.0.5"`, etc.).

- Wrap every bare `https://...` URL in a Rust comment or doc comment in
  angle brackets so they render as autolinks. Five `// Ported from <URL>`
  markers, the install-pipeline link in `tests/real_fs.rs`'s preamble,
  the parallel-disk-usage link in `tests/fakes.rs`'s preamble, and the
  upstream permalinks inside per-test docstrings all use the
  angle-bracketed form now.
