+++ crates/modules-yaml/tests/real_fs.rs
@@ -1,0 +1,130 @@
+//! Pacquet-side tests that exercise behavior branches by writing
+//! `.modules.yaml` files to a real `tempfile::tempdir()` and reading them
+//! back. These cover branches that pnpm only exercises transitively
+//! through install-level integration tests in `pnpm/test/`
+//! (e.g. custom `virtualStoreDir` at
+//! <https://github.com/pnpm/pnpm/blob/1819226b51/pnpm/test/monorepo/index.ts#L1467-L1545>);
+//! the install integration tests are gated on the install pipeline being
+//! ported, so these direct unit tests guard the behavior in the meantime.
+
+use indexmap::IndexSet;
+use pacquet_modules_yaml::{
+    DepPath, Modules, RealApi, read_modules_manifest, write_modules_manifest,
+};
+use pipe_trait::Pipe;
+use pretty_assertions::assert_eq;
+use serde_json::{Value, json};
+use std::{fs, path::Path};
+
+fn manifest_from_json(value: Value) -> Modules {
+    serde_json::from_value(value).expect("deserialize Modules fixture")
+}
+
+/// Reading a manifest whose `virtualStoreDir` is already absolute must
+/// preserve it verbatim, matching upstream
+/// <https://github.com/pnpm/pnpm/blob/1819226b51/installing/modules-yaml/src/index.ts#L66-L70>.
+#[test]
+fn read_preserves_absolute_virtual_store_dir() {
+    let temp_dir = tempfile::tempdir().expect("create temporary directory");
+    let modules_dir = temp_dir.path().join("node_modules");
+    fs::create_dir_all(&modules_dir).expect("create modules dir");
+    let custom_store = temp_dir.path().join("custom-store");
+    let raw = json!({ "virtualStoreDir": &custom_store, "layoutVersion": 5 }).to_string();
+    fs::write(modules_dir.join(".modules.yaml"), raw).expect("write fixture");
+
+    let manifest = modules_dir
+        .pipe_as_ref(read_modules_manifest::<RealApi>)
+        .expect("read manifest")
+        .expect("manifest exists");
+    assert_eq!(Path::new(&manifest.virtual_store_dir), custom_store);
+}
+
+/// `writeModules` sorts `skipped` in place before serializing, matching
+/// upstream
+/// <https://github.com/pnpm/pnpm/blob/1819226b51/installing/modules-yaml/src/index.ts#L117>.
+#[test]
+fn write_sorts_skipped_array() {
+    let temp_dir = tempfile::tempdir().expect("create temporary directory");
+    let modules_dir = temp_dir.path();
+    let manifest = manifest_from_json(json!({
+        "layoutVersion": 5,
+        "skipped": ["zeta", "alpha", "mu"],
+    }));
+
+    write_modules_manifest::<RealApi>(modules_dir, manifest).expect("write manifest");
+    let raw: Value = modules_dir
+        .join(".modules.yaml")
+        .pipe(fs::read_to_string)
+        .expect("read raw .modules.yaml")
+        .pipe_as_ref(serde_json::from_str)
+        .expect("parse raw .modules.yaml");
+    assert_eq!(raw["skipped"], json!(["alpha", "mu", "zeta"]));
+}
+
+/// A null `publicHoistPattern` is removed before serializing because the
+/// YAML writer fails on undefined fields upstream. The behavior matches
+/// <https://github.com/pnpm/pnpm/blob/1819226b51/installing/modules-yaml/src/index.ts#L123-L125>.
+#[test]
+fn write_removes_null_public_hoist_pattern() {
+    let temp_dir = tempfile::tempdir().expect("create temporary directory");
+    let modules_dir = temp_dir.path();
+    let manifest = manifest_from_json(json!({
+        "layoutVersion": 5,
+        "publicHoistPattern": null,
+    }));
+
+    write_modules_manifest::<RealApi>(modules_dir, manifest).expect("write manifest");
+    let raw: Value = modules_dir
+        .join(".modules.yaml")
+        .pipe(fs::read_to_string)
+        .expect("read raw .modules.yaml")
+        .pipe_as_ref(serde_json::from_str)
+        .expect("parse raw .modules.yaml");
+    assert!(
+        raw.get("publicHoistPattern").is_none(),
+        "publicHoistPattern was kept after write: {raw}",
+    );
+}
+
+/// `DepPath` is a transparent newtype around `String`: on the wire it is
+/// indistinguishable from a plain string, so `hoistedAliases` keys and
+/// `ignoredBuilds` elements round-trip through JSON (and YAML) the same
+/// way upstream's `as DepPath`-cast values do.
+#[test]
+fn dep_path_serializes_transparently() {
+    let temp_dir = tempfile::tempdir().expect("create temporary directory");
+    let modules_dir = temp_dir.path();
+    let manifest = manifest_from_json(json!({
+        "layoutVersion": 5,
+        "hoistedAliases": {
+            "/accepts/1.3.7": ["accepts"],
+        },
+        "ignoredBuilds": ["/sharp/0.32.0"],
+        "publicHoistPattern": [],
+    }));
+    assert_eq!(
+        manifest.hoisted_aliases.as_ref().and_then(|m| m.keys().next()),
+        Some(&DepPath::from("/accepts/1.3.7".to_string())),
+    );
+    let expected_ignored: IndexSet<DepPath> =
+        [DepPath::from("/sharp/0.32.0".to_string())].into_iter().collect();
+    assert_eq!(manifest.ignored_builds.as_ref(), Some(&expected_ignored));
+
+    write_modules_manifest::<RealApi>(modules_dir, manifest).expect("write manifest");
+    let raw: Value = modules_dir
+        .join(".modules.yaml")
+        .pipe(fs::read_to_string)
+        .expect("read raw .modules.yaml")
+        .pipe_as_ref(serde_json::from_str)
+        .expect("parse raw .modules.yaml");
+    assert_eq!(
+        raw["hoistedAliases"]["/accepts/1.3.7"],
+        json!(["accepts"]),
+        "DepPath key did not serialize as a plain string",
+    );
+    assert_eq!(
+        raw["ignoredBuilds"],
+        json!(["/sharp/0.32.0"]),
+        "DepPath element did not serialize as a plain string",
+    );
+}
